# feat: v0.4.0 performance optimizations - display-specific rendering