… benches/v0_4_0_performance.rs
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/animation.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/builder.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/linear.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/mod.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/range.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/scale.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/style.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/ticks.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/axes/traits.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/bar.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/curve.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/custom.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/gauge.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/line.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/pie.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/scatter.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/stacked.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/chart/traits.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/dashboard/grid.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/dashboard/layout.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/dashboard/mod.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/dashboard/simple.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/data/aggregation.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/data/bounds.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/data/point.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/data/ring_buffer.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/data/series.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/data/streaming.rs.html
… coverage_report/html/coverage/home/tasuku/com/github/signal-slot/embedded-charts/src/error.rs.html