## Summary
This PR implements the first phase of v0.4.0 performance optimizations, focusing on display-specific rendering optimizations and establishing a comprehensive performance benchmarking infrastructure.

## Changes

### 🚀 Performance Benchmarking Infrastructure
- Created comprehensive benchmarking suite for v0.4.0 performance goals
- Established baseline performance measurements:
  - Line chart rendering (200 points): **20.74 µs**
  - Data series allocation: **634 ns**
  - Point transformation: **166 ns**

### 🖥️ Display-Specific Rendering Optimizations
- Implemented `OptimizedRenderer` trait for display-specific optimizations
- Created three specialized renderers with targeted optimizations:

#### OLED Renderer
- Column-based updates for efficient vertical line rendering
- Batch operations to minimize I2C/SPI transactions
- Performance: **3.25 µs** for 50 vertical lines

#### TFT Renderer  
- DMA-friendly horizontal line batching
- Optimized for row-based memory layouts
- Performance: **8.24 µs** for 50 horizontal lines

#### E-Paper Renderer
- Update region tracking for partial refresh
- Minimizes full-screen refresh operations
- Reduces power consumption and improves responsiveness

### 📁 Module Restructuring
- Reorganized `render` module into proper module structure
- Separated base rendering from optimized implementations
- Maintained backward compatibility with all existing APIs

## Test Plan
- [x] All existing tests pass
- [x] New tests for optimized renderers (6 tests)
- [x] Performance benchmarks establish baselines
- [x] No regression in existing functionality

## Benchmarks
New benchmarks added:
- `baseline_performance` - Establishes v0.4.0 baseline metrics
- `v0_4_0_performance` - Comprehensive performance testing suite
- `display_optimizations` - Display-specific optimization measurements
- `simple_display_bench` - Quick validation benchmarks

## Part of v0.4.0 Release Plan
This PR completes:
- ✅ Phase 2, Week 7-8: Performance optimization infrastructure
- ✅ Display-specific rendering optimizations (OLED, TFT, E-Paper)

Next steps:
- Memory pool management system
- Platform-specific optimizations (ARM Cortex-M, RISC-V, ESP32)
- Target: 20-30% rendering improvement, 15-25% memory reduction

🤖 Generated with [Claude Code](https://claude.ai/code)